+++ components/salsa-macro-rules/src/setup_input_struct.rs
@@ -53,6 +53,8 @@ macro_rules! setup_input_struct {
         // The function used to implement `C::heap_size`.
         heap_size_fn: $($heap_size_fn:path)?,
 
+        page_size: $page_size:ident,
+
         // If `true`, `serialize_fn` and `deserialize_fn` have been provided.
         persist: $persist:tt,
 
@@ -105,6 +107,7 @@ macro_rules! setup_input_struct {
 
                 const PERSIST: bool = $persist;
 
+                type PageSize = $zalsa::$page_size;
                 type Singleton = $zalsa::macro_if! {if $is_singleton {$zalsa::input::Singleton} else {$zalsa::input::NotSingleton}};
 
                 type Struct = $Struct;